pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-orderbus"
version = "0.1.0"
description = "Order Fulfillment POC with Django and Google Pub/Sub"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "psycopg2-binary>=2.9",
    "google-cloud-pubsub>=2.18",
    "httpx>=0.25",
    "python-decouple>=3.8",
]

[tool.setuptools.packages.find]
include = ["orderbus*", "orders*"]
where = ["."]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.1",
    "pytest>=7.4",
    "pytest-django>=4.7",
    "pytest-mock>=3.12",
]

[tool.black]
line-length = 100
target-version = ['py311']
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | \.eggs
  | \.tox
  | __pycache__
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*/migrations/*",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "orderbus.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

